backend/requirements.txt
mysql-connector-python
fastapi[standard]
pydantic
python-dotenv
numpy
pandas
pillow
torch
torchvision